More modular implementation of x86 and x64 Atomic which will allow having that implementation also work on GCC instead of only MSVC, since apparently the GCC 4.6 implementation is also slow.

+++ src/mathicgb/PolyRing.hpp
@@ -596,7 +596,7 @@ private:
 
   coefficient mCharac; // p=mCharac: ring is ZZ/p
   size_t mNumVars;
-  int mNumWeights; // stored as negative of weight vectors
+  size_t mNumWeights; // stored as negative of weight vectors
   size_t mTopIndex;
   size_t mHashIndex; // 1 more than mTopIndex.  Where the has value is stored.
   size_t mMaxMonomialSize;
